refractor some things into containers
