--- src/components/layout/Navbar.js
@@ -1,32 +1,0 @@
-// React
-import React from "react";
-import { Link } from "react-router-dom";
-
-// Redux
-import { connect } from "react-redux";
-
-// Child components
-import SignedInLinks from "./SignedInLinks";
-import SignedOutLinks from "./SignedOutLinks";
-
-const Navbar = props => {
-  const { auth } = props;
-  const links = auth.uid ? <SignedInLinks /> : <SignedOutLinks />;
-
-  return (
-    <nav>
-      <div>
-        <Link to="/">Home</Link>
-        {links}
-      </div>
-    </nav>
-  );
-};
-
-const mapStateToProps = state => {
-  return {
-    auth: state.firebase.auth
-  };
-};
-
-export default connect(mapStateToProps)(Navbar);
