--- src/components/layout/SignedOutLinks.js
@@ -1,20 +1,0 @@
-// React
-import React from "react";
-import { NavLink } from "react-router-dom";
-
-const SignedOutLinks = () => {
-  return (
-    <div>
-      <ul>
-        <li>
-          <NavLink to="/signup">Sign Up</NavLink>
-        </li>
-        <li>
-          <NavLink to="/login">Login</NavLink>
-        </li>
-      </ul>
-    </div>
-  );
-};
-
-export default SignedOutLinks;
